refactor: 리뷰 생성 정리
