+++ src/test/java/com/codeit/duckhu/review/service/ReviewServiceTest.java
@@ -2,6 +2,7 @@ package com.codeit.duckhu.review.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.codeit.duckhu.review.dto.ReviewCreateRequest;
@@ -9,7 +10,9 @@ import com.codeit.duckhu.review.dto.ReviewDto;
 import com.codeit.duckhu.review.entity.Review;
 import com.codeit.duckhu.review.mapper.ReviewMapper;
 import com.codeit.duckhu.review.repository.ReviewRepository;
+import com.codeit.duckhu.review.service.impl.ReviewServiceImpl;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -25,53 +28,59 @@ class ReviewServiceTest {
 
   @Mock
   private ReviewRepository reviewRepository;
-  
+
   @Mock
   private ReviewMapper reviewMapper;
 
   @InjectMocks
   private ReviewServiceImpl reviewService;
-  
+
   /**
    * 서비스 계층 설계
-   * 
-   * 1. ReviewService 인터페이스 생성
-   *    - createReview: 리뷰 생성
-   *    - getReview: ID로 리뷰 조회
-   *    - updateReview: 리뷰 업데이트
-   *    - deleteReview: 리뷰 삭제
-   *    - likeReview: 리뷰 좋아요
-   *    - getAll: 목록 조회(커서 페이지네이션)
+   *
+   * 1. ReviewService
+   * - createReview: 리뷰 생성
+   * - getReview: ID로 리뷰 조회
+   * - updateReview: 리뷰 업데이트
+   * - deleteReview: 리뷰 삭제
+   * - likeReview: 리뷰 좋아요
+   * - getAll: 목록 조회(커서 페이지네이션)
    *
    * 2. DTO 클래스 설계
-   *    - ReviewDto: 응답 데이터
-   *    - ReviewCreateRequest: 생성 요청 데이터
-   *    - ReviewUpdateRequest: 업데이트 요청 데이터
-   *    - ReviewLikeDto: 리뷰 좋아요 데이터
-   * 
-   * 3. 서비스 구현체 (ReviewServiceImpl) 생성
-   *    - 리포지토리를 통한 CRUD 구현
+   * - ReviewDto: 응답 데이터
+   * - ReviewCreateRequest: 생성 요청 데이터
+   * - ReviewUpdateRequest: 업데이트
+   * 요청 데이터 - ReviewLikeDto: 리뷰 좋아요 데이터
+   *
+   * 3. 서비스 구현체 (ReviewServiceImpl) 생성 - 리포지토리를 통한 CRUD 구현
    */
-  
-  @Test
+
+  @Nested
   @DisplayName("리뷰 생성 테스트")
-  void 리뷰_생성() {
+  class CreateReviewTest {
+
+    @Test
+    @DisplayName("유효한 리뷰 데이터로 리뷰 생성 성공")
+    void createReview_withValidData_shouldCreateReview() {
       // Given
       ReviewCreateRequest request = ReviewCreateRequest.builder()
-         .rating(3)
-         .content("볼만해요")
-         .build();
-      
+          .rating(3)
+          .content("볼만해요")
+          .build();
+
       Review review = Review.builder()
           .rating(request.getRating())
           .content(request.getContent())
+          .likeCount(0)
+          .commentCount(0)
+          .likeByMe(false)
           .build();
 
       ReviewDto expectedDto = ReviewDto.builder()
           .content("볼만해요")
           .rating(3)
           .build();
-      
+
       when(reviewRepository.save(any(Review.class))).thenReturn(review);
       when(reviewMapper.toDto(any(Review.class))).thenReturn(expectedDto);
 
@@ -82,5 +91,8 @@ class ReviewServiceTest {
       assertThat(result).isNotNull();
       assertThat(result.getRating()).isEqualTo(3);
       assertThat(result.getContent()).isEqualTo("볼만해요");
+      verify(reviewRepository).save(any(Review.class));
+      verify(reviewMapper).toDto(any(Review.class));
+    }
   }
 }
